dev: review refactor
separation of logic to 2 new components
