+++ src/app/widgets/product-page/reviews/ui/new-review/new-review.component.scss
@@ -1,0 +1,106 @@
+.wrapper {
+    display: flex;
+    flex-direction: column;
+    gap: 12px;
+
+    .form {
+        padding: 16px 12px;
+
+        display: flex;
+        flex-direction: column;
+        gap: 12px;
+
+        box-shadow: var(--medium-shadow);
+
+        .title {
+            font-size: 18px;
+            font-weight: 600;
+
+            user-select: none;
+        }
+        .form-wrapper {
+            input[type=text] {
+                padding: 4px 4px;
+                width: 100%;
+                box-sizing: border-box;
+
+                transition: all 0.2s ease-out;
+
+                outline: none;
+                border: none;
+                border-bottom: 2px solid var(--black-color);
+
+                font-size: 16px;
+                font-family: var(--regular-font);
+            }
+            input[type=text]:focus {
+                border: none;
+                border-bottom: 2px solid var(--light-gray-color);
+            }
+        }
+        .stars-rating {
+            display: flex;
+            gap: 4px;
+
+            user-select: none;
+
+            .star {
+                display: flex;
+                justify-content: center;
+                align-items: center;
+
+                width: 22px;
+                height: 22px;
+
+                font-size: 18px;
+                color: var(--gray-color);
+                text-shadow: 2px 2px 4px rgba(28, 23, 23, 0.1);
+            
+                cursor: pointer;
+                transition: all 0.2s ease-out;
+            }
+
+            .active-star {
+                color: var(--yellow-color);
+            }
+        }
+    }        
+    .send-review {
+        width: 250px;
+        padding: 8px 22px;
+        box-sizing: border-box;
+
+        display: flex;
+        justify-content: center;
+        align-items: center;
+
+        border: none;
+        background-color: var(--yellow-color);
+
+        p {
+            width: max-content;
+
+            font-size: 16px;
+            font-weight: 600;
+            color: var(--black-color);
+
+            user-select: none;
+        }
+    }
+    .send-review:hover {
+        box-shadow: var(--small-shadow);
+    }
+    .send-review:active {
+        box-shadow: var(--small-shadow);
+        transition: 0.2s;
+        transform: scale(0.98);
+    }
+}
+
+@media screen and (max-width: 512px) {
+    .wrapper {
+        .send-review {
+            width: 100%;
+        }
+    }
+}
