+++ src/app/widgets/product-page/reviews/ui/new-review/new-review.component.ts
@@ -1,0 +1,47 @@
+import { Component, inject, input } from '@angular/core';
+import { FormBuilder, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
+import { ActivatedRoute } from '@angular/router';
+import { ReviewsService } from '../../api/reviews.service';
+
+@Component({
+  selector: 'app-new-review',
+  standalone: true,
+  imports: [ ReactiveFormsModule ],
+  templateUrl: './new-review.component.html',
+  styleUrl: './new-review.component.scss'
+})
+export class NewReviewComponent {
+  private router: ActivatedRoute = inject(ActivatedRoute);
+  private reviewsService: ReviewsService = inject(ReviewsService);
+  private fb: FormBuilder = inject(FormBuilder);
+
+  productId = input.required();
+
+  sendingReview = false;
+
+  customerReview: FormGroup = this.fb.group({
+    customerReviewText: ['', Validators.required],
+  });
+
+  sendReview() {
+    if (this.customerReview.valid && !this.sendingReview) {
+      this.sendingReview = true;
+
+      this.reviewsService.sendNewReview(
+        new Date().toISOString(),
+        '1',
+        'джоу доу',
+        this.productId.toString(),
+        this.customerReview.get('customerReviewText')?.value,
+        5,
+      ).subscribe((res: any) => {
+        window.alert('ваш відгук успрішно надіслано!\nвін зв\'явиться на сайті щойно пройде модеріцію')
+        this.sendingReview = false;
+        this.reviewsService.getReviewsData(this.productId.toString());
+        this.customerReview.reset();
+      })
+    } else {
+      window.alert('будь ласка, заповніть поле відгуку')
+    }
+  }
+}
