+++ src/app/widgets/product-page/reviews/ui/product-reviews/product-reviews.component.scss
@@ -1,0 +1,147 @@
+.wrapper {
+    display: flex;
+    flex-direction: column;
+    gap: 24px;
+
+    .reviews-data {
+        display: flex;
+        flex-direction: row;
+        gap: 24px;
+
+        .reviews-number {
+            p {
+                font-size: 22px;
+                font-weight: 400;
+            }
+        }
+        .product-rating {
+            display: flex;
+            gap: 4px;
+
+            user-select: none;
+
+            .star {
+                display: flex;
+                justify-content: center;
+                align-items: center;
+
+                width: 22px;
+                height: 22px;
+
+                font-size: 18px;
+                color: var(--yellow-color);
+                text-shadow: 2px 2px 4px rgba(28, 23, 23, 0.1);                      
+            }
+        }
+    }
+
+    .reviews {
+        display: flex;
+        flex-direction: column;
+        gap: 24px;
+
+        max-height: 300px;
+        overflow-y: scroll;
+
+        box-shadow: var(--medium-shadow);
+        
+        .review {
+            padding: 0px 11px;
+            display: flex;
+            gap: 18px;
+
+            background-color: var(--white-color);
+            box-shadow: var(--medium-shadow);
+
+            .user-data {
+                padding: 11px 0px;
+
+                display: flex;
+                flex-direction: column;
+                align-items: center;
+                gap: 11px;
+
+                max-width: 88px;
+
+                .picture {
+                    width: 44px;
+                    height: 44px;
+                    background-color: var(--yellow-color);
+
+                    * {
+                        width: 100%;
+                        height: 100%;
+                        object-fit: cover;
+                    }
+                    div {
+                        display: flex;
+                        justify-content: center;
+                        align-items: center;
+
+                        font-weight: 700;
+                        user-select: none;
+                    }
+                }
+                .name {
+                    width: 100%;
+
+                    p {
+                        font-size: 16px;
+                        text-align: center;
+
+                        // to make text overflow with ellipsis in 2 lines
+                        display: -webkit-box;
+                        line-clamp: 2;
+                        -webkit-line-clamp: 2;
+                        -webkit-box-orient: vertical;
+                        overflow: hidden;
+                        text-overflow: ellipsis;
+                        max-height: 3em;
+                        line-height: 1.5em;
+                    }
+                }
+            }
+            .review-data {
+                padding: 11px 0px;
+
+                display: flex;
+                flex-direction: column;
+                gap: 11px;
+
+                .review-rating {
+                    display: flex;
+                    gap: 4px;
+
+                    user-select: none;
+
+                    .star {
+                        display: flex;
+                        justify-content: center;
+                        align-items: center;
+
+                        width: 22px;
+                        height: 22px;
+
+                        font-size: 18px;
+                        color: var(--yellow-color);
+                        text-shadow: 2px 2px 4px rgba(28, 23, 23, 0.1);
+                    }
+                }
+                .review-text {
+                    p {
+                        font-size: 18px;
+                    }
+                }
+            }
+        }
+    }
+}
+
+@media screen and (max-width: 512px) {
+    .wrapper {
+        .reviews-data {
+            justify-content: space-between;
+            gap: 12px;
+        }
+    }    
+}
